+++ lib/recording_studio_api/resource_operation_context.rb
@@ -14,8 +14,28 @@ module RecordingStudioApi
     :params,
     :request_params,
     :scoped_recordings,
-    :parent_recording
+    :parent_recording,
+    :idempotency_key
   ) do
+    def initialize( # rubocop:disable Metrics/ParameterLists
+      recording:,
+      recordable_type:,
+      resource_name:,
+      api_client:,
+      credential:,
+      access_recording:,
+      access_grant:,
+      root_recording:,
+      api_version:,
+      params:,
+      request_params:,
+      scoped_recordings:,
+      parent_recording:,
+      idempotency_key: nil
+    )
+      super
+    end
+
     def api_key
       api_client&.api_key.presence || "public"
     end